Fix `compute_payment()` and test all han/fu combinations
The previous version of `compute_payment()` had several bugs pertaining
to hands with a han score below 5:

  - It was off by 100 for many dealer ron scores, due to rounding

  - It did not support values of fu > 100

  - It did not properly identify hands as "Mangan"

(It also wrongly returned 1400 for values of 2 han, 20 fu and non-dealer
ron, although this is mostly theoretical, as such a hand cannot exist
with the current rules.)

There is actually no reason to use a lookup table instead of calculating
the score directly, which is what we do here.  Testing was also modified
to automatically test every possible han/fu combination.